Move checkPaymentMethodIsAvailable to a separate file
